requirements.txt
Adafruit-DHT==1.4.0
numpy==1.22.3
PyBluez==0.23
pycryptodome==3.14.1
requests==2.27.1
RPi.GPIO==0.7.1
# setuptools > 58 break installation of PyBluez
setuptools<58